@@ -1,12 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net7.0-windows</TargetFramework>
+    <TargetFramework>net10.0-windows10.0.26100</TargetFramework>
     <UseWindowsForms>true</UseWindowsForms>
-    <Version>1.2.2</Version>
-    <AssemblyVersion>1.2.2</AssemblyVersion>
-    <FileVersion>1.2.2</FileVersion>
-    <Copyright>Copyright © Joshua Kempner 2008-2022</Copyright>
+    <Version>1.3.0</Version>
+    <AssemblyVersion>1.3.0</AssemblyVersion>
+    <FileVersion>1.3.0</FileVersion>
+    <Copyright>Copyright © Joshua Kempner 2008-2026</Copyright>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="..\cosmo\Cosmo.csproj" />
