@@ -1,12 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <Version>1.2.2</Version>
-    <AssemblyVersion>1.2.2</AssemblyVersion>
-    <FileVersion>1.2.2</FileVersion>
-    <Copyright>Copyright © Joshua Kempner 2008-2022</Copyright>
+    <TargetFramework>net10.0</TargetFramework>
+    <Version>1.3.0</Version>
+    <AssemblyVersion>1.3.0</AssemblyVersion>
+    <FileVersion>1.3.0</FileVersion>
+    <Copyright>Copyright © Joshua Kempner 2008-2026</Copyright>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="reactiveui" Version="18.4.34" />
+    <PackageReference Include="reactiveui" Version="22.3.1" />
   </ItemGroup>
 </Project>
